get crawl out of $ICLOUD
I don't actually play this much anymore, but if I do it would probably be on a plane, and I don't want it trying to sync anything when I'm on plane wifi, and I'm not really going to pick a game back up on my desktop at home.

I prefer playing online via web when possible anyhow for preserving scores/stats.